chore: remove unnecesarry auth squats
