+++ pkg/ingester/otlp/ingest_handler_test.go
@@ -1,17 +1,24 @@
 package otlp
 
 import (
+	"bytes"
 	"context"
 	"flag"
+	"net/http"
+	"net/http/httptest"
 	"os"
 	"sort"
 	"strings"
 	"testing"
 
 	"github.com/grafana/dskit/server"
+	"github.com/grafana/dskit/user"
+	"github.com/klauspost/compress/gzip"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
+	"google.golang.org/protobuf/proto"
+
 	v1experimental2 "go.opentelemetry.io/proto/otlp/collector/profiles/v1development"
 	v1 "go.opentelemetry.io/proto/otlp/common/v1"
 	v1experimental "go.opentelemetry.io/proto/otlp/profiles/v1development"
@@ -20,8 +27,10 @@ import (
 	"github.com/grafana/pyroscope/pkg/distributor/model"
 	phlaremodel "github.com/grafana/pyroscope/pkg/model"
 	"github.com/grafana/pyroscope/pkg/og/convert/pprof/strprofile"
+	"github.com/grafana/pyroscope/pkg/tenant"
 	"github.com/grafana/pyroscope/pkg/test"
 	"github.com/grafana/pyroscope/pkg/test/mocks/mockotlp"
+	"github.com/grafana/pyroscope/pkg/util"
 )
 
 func TestGetServiceNameFromAttributes(t *testing.T) {
@@ -344,8 +353,8 @@ func TestConversion(t *testing.T) {
 						}}}}},
 				Dictionary: &b.dictionary}
 			logger := test.NewTestingLogger(t)
-			h := NewOTLPIngestHandler(testConfig(), svc, logger, false)
-			_, err := h.Export(context.Background(), req)
+			h := NewOTLPIngestHandler(testConfig(), svc, logger)
+			_, err := h.Export(user.InjectOrgID(context.Background(), tenant.DefaultTenantID), req)
 
 			if td.expectedError == "" {
 				require.NoError(t, err)
@@ -443,8 +452,8 @@ func TestSampleAttributes(t *testing.T) {
 				}}}}},
 		Dictionary: &otlpb.dictionary}
 	logger := test.NewTestingLogger(t)
-	h := NewOTLPIngestHandler(testConfig(), svc, logger, false)
-	_, err := h.Export(context.Background(), req)
+	h := NewOTLPIngestHandler(testConfig(), svc, logger)
+	_, err := h.Export(user.InjectOrgID(context.Background(), tenant.DefaultTenantID), req)
 	assert.NoError(t, err)
 	require.Equal(t, 1, len(profiles))
 	require.Equal(t, 1, len(profiles[0].Series))
@@ -609,8 +618,8 @@ func TestDifferentServiceNames(t *testing.T) {
 		Dictionary: &otlpb.dictionary}
 
 	logger := test.NewTestingLogger(t)
-	h := NewOTLPIngestHandler(testConfig(), svc, logger, false)
-	_, err := h.Export(context.Background(), req)
+	h := NewOTLPIngestHandler(testConfig(), svc, logger)
+	_, err := h.Export(user.InjectOrgID(context.Background(), tenant.DefaultTenantID), req)
 	require.NoError(t, err)
 
 	require.Equal(t, 1, len(profiles))
@@ -671,4 +680,166 @@ func testConfig() server.Config {
 	fs := flag.NewFlagSet("test", flag.PanicOnError)
 	cfg.RegisterFlags(fs)
 	return cfg
+}
+
+// createValidOTLPRequest creates a minimal valid OTLP profile export request for testing
+func createValidOTLPRequest() *v1experimental2.ExportProfilesServiceRequest {
+	b := new(otlpbuilder)
+	b.dictionary.MappingTable = []*v1experimental.Mapping{{
+		MemoryStart:      0x1000,
+		MemoryLimit:      0x2000,
+		FilenameStrindex: b.addstr("test.so"),
+	}}
+	b.dictionary.LocationTable = []*v1experimental.Location{{
+		MappingIndex: 0,
+		Address:      0x1100,
+	}}
+	b.dictionary.StackTable = []*v1experimental.Stack{{
+		LocationIndices: []int32{0},
+	}}
+	b.profile.SampleType = &v1experimental.ValueType{
+		TypeStrindex: b.addstr("samples"),
+		UnitStrindex: b.addstr("count"),
+	}
+	b.profile.Sample = []*v1experimental.Sample{{
+		StackIndex: 0,
+		Values:     []int64{100},
+	}}
+	b.profile.TimeUnixNano = 1234567890
+
+	return &v1experimental2.ExportProfilesServiceRequest{
+		ResourceProfiles: []*v1experimental.ResourceProfiles{{
+			ScopeProfiles: []*v1experimental.ScopeProfiles{{
+				Profiles: []*v1experimental.Profile{&b.profile},
+			}},
+		}},
+		Dictionary: &b.dictionary,
+	}
+}
+
+func TestHTTPRequestWithJSONAndTenantAccepted(t *testing.T) {
+	svc := mockotlp.NewMockPushService(t)
+	var capturedTenantID string
+	svc.On("PushBatch", mock.Anything, mock.Anything).Run(func(args mock.Arguments) {
+		ctx := args.Get(0).(context.Context)
+		tenantID, err := tenant.ExtractTenantIDFromContext(ctx)
+		require.NoError(t, err)
+		capturedTenantID = tenantID
+	}).Return(nil, nil)
+
+	logger := test.NewTestingLogger(t)
+	h := NewOTLPIngestHandler(testConfig(), svc, logger)
+
+	jsonRequest := `{
+		"resourceProfiles": [{
+			"scopeProfiles": [{
+				"profiles": [{
+					"sampleType": {"typeStrindex": 0, "unitStrindex": 1},
+					"sample": [{"stackIndex": 0, "values": [100]}],
+					"timeUnixNano": "1234567890"
+				}]
+			}]
+		}],
+		"dictionary": {
+			"stringTable": ["samples", "count", "test.so"],
+			"mappingTable": [{"memoryStart": "4096", "memoryLimit": "8192", "filenameStrindex": 2}],
+			"locationTable": [{"mappingIndex": 0, "address": "4352"}],
+			"stackTable": [{"locationIndices": [0]}]
+		}
+	}`
+
+	httpReq := httptest.NewRequest("POST", "/otlp/v1/profiles", bytes.NewReader([]byte(jsonRequest)))
+	httpReq.Header.Set("Content-Type", "application/json")
+	httpReq.Header.Set(user.OrgIDHeaderName, "json-tenant")
+
+	w := httptest.NewRecorder()
+	util.AuthenticateUser(true).Wrap(h).ServeHTTP(w, httpReq)
+
+	assert.Equal(t, http.StatusOK, w.Code)
+	assert.Equal(t, "json-tenant", capturedTenantID)
+}
+
+func TestHTTPRequestWithGzipCompression(t *testing.T) {
+	svc := mockotlp.NewMockPushService(t)
+	var capturedTenantID string
+	svc.On("PushBatch", mock.Anything, mock.Anything).Run(func(args mock.Arguments) {
+		ctx := args.Get(0).(context.Context)
+		tenantID, err := tenant.ExtractTenantIDFromContext(ctx)
+		require.NoError(t, err)
+		capturedTenantID = tenantID
+	}).Return(nil, nil)
+
+	logger := test.NewTestingLogger(t)
+	h := NewOTLPIngestHandler(testConfig(), svc, logger)
+
+	req := createValidOTLPRequest()
+	reqBytes, err := proto.Marshal(req)
+	require.NoError(t, err)
+
+	var gzipBuf bytes.Buffer
+	gzipWriter := gzip.NewWriter(&gzipBuf)
+	_, err = gzipWriter.Write(reqBytes)
+	require.NoError(t, err)
+	err = gzipWriter.Close()
+	require.NoError(t, err)
+
+	httpReq := httptest.NewRequest("POST", "/otlp/v1/profiles", bytes.NewReader(gzipBuf.Bytes()))
+	httpReq.Header.Set("Content-Type", "application/x-protobuf")
+	httpReq.Header.Set("Content-Encoding", "gzip")
+
+	w := httptest.NewRecorder()
+	util.AuthenticateUser(false).Wrap(h).ServeHTTP(w, httpReq)
+
+	assert.Equal(t, http.StatusOK, w.Code)
+	assert.Equal(t, tenant.DefaultTenantID, capturedTenantID)
+}
+
+func TestHTTPRequestWithGzipCompressionAndJSON(t *testing.T) {
+	svc := mockotlp.NewMockPushService(t)
+	var capturedTenantID string
+	svc.On("PushBatch", mock.Anything, mock.Anything).Run(func(args mock.Arguments) {
+		ctx := args.Get(0).(context.Context)
+		tenantID, err := tenant.ExtractTenantIDFromContext(ctx)
+		require.NoError(t, err)
+		capturedTenantID = tenantID
+	}).Return(nil, nil)
+
+	logger := test.NewTestingLogger(t)
+	h := NewOTLPIngestHandler(testConfig(), svc, logger)
+
+	jsonRequest := `{
+		"resourceProfiles": [{
+			"scopeProfiles": [{
+				"profiles": [{
+					"sampleType": {"typeStrindex": 0, "unitStrindex": 1},
+					"sample": [{"stackIndex": 0, "values": [100]}],
+					"timeUnixNano": "1234567890"
+				}]
+			}]
+		}],
+		"dictionary": {
+			"stringTable": ["samples", "count", "test.so"],
+			"mappingTable": [{"memoryStart": "4096", "memoryLimit": "8192", "filenameStrindex": 2}],
+			"locationTable": [{"mappingIndex": 0, "address": "4352"}],
+			"stackTable": [{"locationIndices": [0]}]
+		}
+	}`
+
+	var gzipBuf bytes.Buffer
+	gzipWriter := gzip.NewWriter(&gzipBuf)
+	_, err := gzipWriter.Write([]byte(jsonRequest))
+	require.NoError(t, err)
+	err = gzipWriter.Close()
+	require.NoError(t, err)
+
+	httpReq := httptest.NewRequest("POST", "/otlp/v1/profiles", bytes.NewReader(gzipBuf.Bytes()))
+	httpReq.Header.Set("Content-Type", "application/json")
+	httpReq.Header.Set("Content-Encoding", "gzip")
+	httpReq.Header.Set(user.OrgIDHeaderName, "gzip-json-tenant")
+
+	w := httptest.NewRecorder()
+	util.AuthenticateUser(true).Wrap(h).ServeHTTP(w, httpReq)
+
+	assert.Equal(t, http.StatusOK, w.Code)
+	assert.Equal(t, "gzip-json-tenant", capturedTenantID)
 }
