+++ pkg/pyroscope/modules.go
@@ -324,7 +324,7 @@ func (f *Pyroscope) initDistributor() (services.Service, error) {
 	if err != nil {
 		return nil, err
 	}
-	f.API.RegisterDistributor(d, f.Overrides, f.Cfg.MultitenancyEnabled, f.Cfg.Server)
+	f.API.RegisterDistributor(d, f.Overrides, f.Cfg.Server)
 	return d, nil
 }
